requirements.txt
altair
cron_descriptor
streamlit==1.29.0
snowflake-snowpark-python[pandas]==1.6.1
pandas==1.5.3
plotly
cffi==1.15.1
st-pages